Stage 3, elo logic, calculations, ENUMs, new Exceptions

+++ src/main/java/org/training/dto/player/PlayerResponse.java
@@ -16,7 +16,7 @@ public class PlayerResponse {
 
     private String email;
 
-    private Integer eloRating;
+    private Double eloRating;
 
     private LocalDate registeredAt;
